Crate abc_algorithm

Reexports

pub use StopCriterion::*;

Structs

Config
Provide means to adjust the main paramters driving the algorithm.

Solution
Groups information relevant to a solution.

Enums

StopCriterion
A set of possible criteria to decide when to stop the optimization.