Enum abc_algorithm::StopCriterion

pub enum StopCriterion {
    MaxFunctionEvaluations(usize),
    MaxIterations(usize),
    TargetValue(f64, f64),
}

A set of possible criteria to decide when to stop the optimization.

Variants

MaxFunctionEvaluations

Stop if a specified number of function evaluations has been reached.

MaxIterations

Stop if a specified number of iterations has been reached.

TargetValue

Stop if a certain target value range has been reached.

The first value represents the desired target value, while the second one specifies an absolute error which is additionally applied to both directions.