Struct abc_algorithm::Config

pub struct Config {
    // some fields omitted
}

Provide means to adjust the main paramters driving the algorithm.

Methods

impl Config

fn new() -> Config

Create a new Config which describes the behavior of the ABC algorithm.

By default, this configuration provides sane defaults, i.e., it uses 20 food sources, a limit of 100 trials per food source

fn foods(&mut self, foods: usize) -> &mut Config

Set the number of food sources.

fn max_trials(&mut self, max_trials: usize) -> &mut Config

Set the number of maximum trials per food source, bevore abondining the source.

fn stop(&mut self, criterion: StopCriterion) -> &mut Config

Enable the supplied stop criterion for future optimizations.

fn optimize<F>(&self, dimensions: usize, low: f64, high: f64, objective: F) -> Solution where F: Fn(&[f64]) -> f64

Optimize the given objective function by using the ABC algorithm with the specified parameters.